+++ src/contractTest/java/uk/gov/hmcts/reform/wataskmanagementapi/provider/TaskManagementGetTaskByIdPactTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.wataskmanagementapi.provider.service;
+package uk.gov.hmcts.reform.wataskmanagementapi.provider;
 
 import au.com.dius.pact.provider.junit5.PactVerificationContext;
 import au.com.dius.pact.provider.junit5.PactVerificationInvocationContextProvider;
@@ -8,27 +8,30 @@ import au.com.dius.pact.provider.junitsupport.State;
 import au.com.dius.pact.provider.junitsupport.loader.PactBroker;
 import au.com.dius.pact.provider.junitsupport.loader.VersionSelector;
 import au.com.dius.pact.provider.spring.junit5.MockMvcTestTarget;
-import org.json.JSONException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.TestTemplate;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.mockito.Mock;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.hmcts.reform.wataskmanagementapi.auth.access.AccessControlService;
+import uk.gov.hmcts.reform.wataskmanagementapi.auth.access.entities.AccessControlResponse;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.TaskController;
 import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.task.Task;
+import uk.gov.hmcts.reform.wataskmanagementapi.provider.service.TaskManagementProviderTestConfiguration;
 import uk.gov.hmcts.reform.wataskmanagementapi.services.CamundaService;
 
 import java.time.ZonedDateTime;
 
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 
 @ExtendWith(SpringExtension.class)
 @Provider("wa_task_management_api_get_task_by_id")
-//@PactFolder("target/pacts")
+//@PactFolder("pacts")
 @PactBroker(scheme = "${PACT_BROKER_SCHEME:https}",
     host = "${PACT_BROKER_URL:pact-broker.platform.hmcts.net}",
     port = "${PACT_BROKER_PORT:443}", consumerVersionSelectors = {
@@ -37,19 +40,20 @@ import static org.mockito.Mockito.when;
 @IgnoreNoPactsToVerify
 public class TaskManagementGetTaskByIdPactTest {
 
-    @Autowired
+    @Mock
     private AccessControlService accessControlService;
 
-    @Autowired
+    @Mock
     private CamundaService camundaService;
 
     @TestTemplate
     @ExtendWith(PactVerificationInvocationContextProvider.class)
     void pactVerificationTestTemplate(PactVerificationContext context) {
         if (context != null) {
-           // context.verifyInteraction();
+            context.verifyInteraction();
         }
     }
+
 
     @BeforeEach
     void beforeCreate(PactVerificationContext context) {
@@ -65,15 +69,17 @@ public class TaskManagementGetTaskByIdPactTest {
     }
 
     @State({"appropriate task is returned"})
-    public void getTaskById() throws JSONException {
-
+    public void getTaskById() {
+        setInitMockTask();
     }
 
     private void setInitMockTask() {
+        AccessControlResponse accessControlResponse = mock((AccessControlResponse.class));
         when(camundaService.getTask(any(),any(),any())).thenReturn(createTask());
+        when(accessControlService.getRoles(anyString())).thenReturn(accessControlResponse);
     }
 
-    public Task createTask() {
+    private Task createTask() {
         return new Task("4d4b6fgh-c91f-433f-92ac-e456ae34f72a",
                          "Jake",
                          "ReviewTheAppeal",
