+++ src/contractTest/java/uk/gov/hmcts/reform/wataskmanagementapi/provider/TaskManagementGetTaskBySearchForCompletablePactTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.wataskmanagementapi.provider.service;
+package uk.gov.hmcts.reform.wataskmanagementapi.provider;
 
 import au.com.dius.pact.provider.junit5.PactVerificationContext;
 import au.com.dius.pact.provider.junit5.PactVerificationInvocationContextProvider;
@@ -15,8 +15,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.hmcts.reform.wataskmanagementapi.auth.access.AccessControlService;
+import uk.gov.hmcts.reform.wataskmanagementapi.auth.access.entities.AccessControlResponse;
 import uk.gov.hmcts.reform.wataskmanagementapi.controllers.TaskController;
 import uk.gov.hmcts.reform.wataskmanagementapi.domain.entities.task.Task;
+import uk.gov.hmcts.reform.wataskmanagementapi.provider.service.TaskManagementProviderTestConfiguration;
 import uk.gov.hmcts.reform.wataskmanagementapi.services.CamundaService;
 
 import java.time.ZonedDateTime;
@@ -24,17 +26,19 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 
 @ExtendWith(SpringExtension.class)
-@Provider("wa_task_management_api_get_task_by_id")
+@Provider("wa_task_management_api_search_completable")
 @PactBroker(scheme = "${PACT_BROKER_SCHEME:http}",
     host = "${PACT_BROKER_URL:localhost}", port = "${PACT_BROKER_PORT:9292}", consumerVersionSelectors = {
     @VersionSelector(tag = "master")})
 @Import(TaskManagementProviderTestConfiguration.class)
 @IgnoreNoPactsToVerify
-public class TaskManagementGetTaskBySearchCriteriaPactTest {
+public class TaskManagementGetTaskBySearchForCompletablePactTest {
 
     @Autowired
     private AccessControlService accessControlService;
@@ -46,7 +50,7 @@ public class TaskManagementGetTaskBySearchCriteriaPactTest {
     @ExtendWith(PactVerificationInvocationContextProvider.class)
     void pactVerificationTestTemplate(PactVerificationContext context) {
         if (context != null) {
-           // context.verifyInteraction();
+            context.verifyInteraction();
         }
     }
 
@@ -63,59 +67,63 @@ public class TaskManagementGetTaskBySearchCriteriaPactTest {
 
     }
 
-    @State({"appropriate tasks are returned by criteria"})
+
+    @State({"appropriate tasks are returned by search for completable"})
     public void getTaskByCriteria() {
+        setInitMockForSearchByCompletableTask();
     }
 
-    private void setInitMockForsearchTask() {
+    private void setInitMockForSearchByCompletableTask() {
+        AccessControlResponse accessControlResponse = mock((AccessControlResponse.class));
+        when(accessControlService.getRoles(anyString())).thenReturn(accessControlResponse);
         when(camundaService.searchWithCriteria(any(),any(),any())).thenReturn(createTasks());
     }
 
     public List<Task> createTasks() {
         var tasks = new ArrayList<Task>();
         var taskOne =  new Task("4d4b6fgh-c91f-433f-92ac-e456ae34f72a",
-                                              "Jake",
-                                              "ReviewTheAppeal",
-                                              "unconfigured",
-                                              "SELF",
-                                              "PRIVATE",
-                                              "task name",
-                                              ZonedDateTime.now(),
-                                              ZonedDateTime.now(),
-                                              "Mark Alistair",
-                                              true,
-                                              "Time extension",
-                                              "IA",
-                                              "1",
-                                              "765324",
-                                              "Newcastle",
-                                              "Asylum",
-                                              "4d4b3a4e-c91f-433f-92ac-e456ae34f72a",
-                                              "processApplication",
-                                              "Bob Smith",
-                                              true);
+                                "Jake",
+                                "ReviewTheAppeal",
+                                "unconfigured",
+                                "SELF",
+                                "PRIVATE",
+                                "task name",
+                                ZonedDateTime.now(),
+                                ZonedDateTime.now(),
+                                "Mark Alistair",
+                                true,
+                                "Time extension",
+                                "IA",
+                                "1",
+                                "765324",
+                                "Newcastle",
+                                "Asylum",
+                                "4d4b3a4e-c91f-433f-92ac-e456ae34f72a",
+                                "processApplication",
+                                "Bob Smith",
+                                true);
 
         var taskTwo =  new Task("4d4b6fgh-cc1f-433f-92ac-e456aed4f72a",
-                                              "Megan",
-                                              "ReviewTheAppeal",
-                                              "unconfigured",
-                                              "SELF",
-                                              "PRIVATE",
-                                              "task name",
-                                              ZonedDateTime.now(),
-                                              ZonedDateTime.now(),
-                                              "Jean Pierre",
-                                              true,
-                                              "Time extension",
-                                              "IA",
-                                              "1",
-                                              "766524",
-                                              "Newcastle",
-                                              "Asylum",
-                                              "4d4b3a4e-c9df-43sf-92ac-e456ee34fe2a",
-                                              "processApplication",
-                                              "Bob Smith",
-                                              true);
+                                "Megan",
+                                "ReviewTheAppeal",
+                                "unconfigured",
+                                "SELF",
+                                "PRIVATE",
+                                "task name",
+                                ZonedDateTime.now(),
+                                ZonedDateTime.now(),
+                                "Jean Pierre",
+                                true,
+                                "Time extension",
+                                "IA",
+                                "1",
+                                "766524",
+                                "Newcastle",
+                                "Asylum",
+                                "4d4b3a4e-c9df-43sf-92ac-e456ee34fe2a",
+                                "processApplication",
+                                "Bob Smith",
+                                true);
 
         tasks.add(taskOne);
         tasks.add(taskTwo);
