+++ src/contractTest/resources/logback-test.xml
@@ -6,7 +6,7 @@
             </pattern>
         </encoder>
     </appender>
-    <root level="error">
+    <root level="debug">
         <appender-ref ref="STDOUT"/>
     </root>
     <logger name="au.com.dius.pact" level="debug"/>
